feat: add func to play round with values
Added a function with a conditional if. The function will return a
result based on the conditions. The draw option return has a replayRound
function call but the function is yet to be written, so script will
return an error. Also added a new variable for storing a user entered
value which is used in the aforementioned function. A console.log line
has been added to check the outputs of the function.